@@ -733,8 +733,8 @@ class DataHandler:
 
 
 import multiprocessing as mp
-from PySide2 import QtCore
-from PySide2.QtWidgets import QMenuBar, QGridLayout, QApplication, QWidget, QPlainTextEdit, QMainWindow
+from PySide6 import QtCore
+from PySide6.QtWidgets import QMenuBar, QGridLayout, QApplication, QWidget, QPlainTextEdit, QMainWindow
 import sys
 
 class GuiMain(QMainWindow):
