@@ -1,6 +1,6 @@
 import sys, os
-from PySide2 import QtCore
-from PySide2.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QPushButton, QFileDialog
+from PySide6 import QtCore
+from PySide6.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QPushButton, QFileDialog
 
 from pyqtgraph.parametertree import Parameter
 from pyecog2.ui_elements.pyecogParameterTree import PyecogParameterTree
