@@ -1,6 +1,6 @@
 import sys
-from PySide2 import QtGui, QtCore
-from PySide2.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QTextBrowser, QPushButton, QFileDialog
+from PySide6 import QtGui, QtCore
+from PySide6.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QTextBrowser, QPushButton, QFileDialog
 
 from pyecog2.ProjectClass import Animal
 # from ..ProjectClass import Animal
