@@ -1,13 +1,13 @@
 # PyQt5 Video player
 import os
 os.environ['QT_MULTIMEDIA_PREFERRED_PLUGINS'] = 'windowsmediafoundation'
-from PySide2.QtCore import QDir, Qt, QUrl, Signal
-from PySide2.QtMultimedia import QMediaContent, QMediaPlayer
-from PySide2.QtMultimediaWidgets import QVideoWidget
-from PySide2.QtWidgets import (QApplication, QFileDialog, QHBoxLayout, QLabel, QInputDialog,
+from PySide6.QtCore import QDir, Qt, QUrl, Signal
+from PySide6.QtMultimedia import QMediaPlayer
+from PySide6.QtMultimediaWidgets import QVideoWidget
+from PySide6.QtWidgets import (QApplication, QFileDialog, QHBoxLayout, QLabel, QInputDialog,
         QPushButton, QSizePolicy, QSlider, QStyle, QVBoxLayout, QWidget)
-from PySide2.QtWidgets import QMainWindow,QWidget, QPushButton, QAction
-from PySide2.QtGui import QIcon
+from PySide6.QtWidgets import QMainWindow,QWidget, QPushButton
+from PySide6.QtGui import QIcon
 import sys
 import time
 import numpy as np
@@ -32,7 +32,7 @@ class VideoWindow(QWidget):
         self.position_on_new_file = 0
         self.duration = -1
         self.waiting_for_file = False
-        self.media_state_before_file_transition = self.mediaPlayer.state()
+        self.media_state_before_file_transition = self.mediaPlayer.playbackState
         self.video_time_offset = 0.0
 
         self.play_icon = QIcon(play_icon_file)
@@ -73,11 +73,11 @@ class VideoWindow(QWidget):
         self.setLayout(layout)
 
         self.mediaPlayer.setVideoOutput(videoWidget)
-        self.mediaPlayer.stateChanged.connect(self.mediaStateChanged)
+        self.mediaPlayer.playbackStateChanged.connect(self.mediaStateChanged)
         self.mediaPlayer.positionChanged.connect(self.positionChanged)
         self.mediaPlayer.durationChanged.connect(self.durationChanged)
         self.mediaPlayer.mediaStatusChanged.connect(self.mediaStatusChanged)
-        self.mediaPlayer.error.connect(self.handleError)
+        self.mediaPlayer.errorOccurred.connect(self.handleError)
         self.mediaPlayer.setNotifyInterval(40) # 25 fps
 
         # Connect main model time changes to videa seek
@@ -96,7 +96,7 @@ class VideoWindow(QWidget):
             self.current_file = self.main_model.project.current_animal.video_files[0]
             self.current_time_range = [self.main_model.project.current_animal.video_init_time[0],
                                    self.main_model.project.current_animal.video_init_time[0] + self.main_model.project.current_animal.video_duration[0]]
-            self.mediaPlayer.setMedia(QMediaContent(QUrl.fromLocalFile(self.current_file)))
+            self.mediaPlayer.setMedia(QUrl.fromLocalFile(self.current_file))
             self.playButton.setEnabled(True)
         else:
             self.current_file = ''
@@ -118,15 +118,14 @@ class VideoWindow(QWidget):
                 QDir.homePath())
 
         if fileName != '':
-            self.mediaPlayer.setMedia(
-                    QMediaContent(QUrl.fromLocalFile(fileName)))
+            self.mediaPlayer.setMedia(QUrl.fromLocalFile(fileName))
             self.playButton.setEnabled(True)
 
     def exitCall(self):
         sys.exit(app.exec_())
 
     def play(self):
-        if self.mediaPlayer.state() == QMediaPlayer.PlayingState:
+        if self.mediaPlayer.playbackState == QMediaPlayer.PlayingState:
             self.mediaPlayer.pause()
             logger.info("Video player: pausing")
         else:
@@ -135,7 +134,7 @@ class VideoWindow(QWidget):
             logger.info("Video player: playing")
 
     def mediaStateChanged(self, state):
-        if self.mediaPlayer.state() == QMediaPlayer.PlayingState:
+        if self.mediaPlayer.playbackState == QMediaPlayer.PlayingState:
             self.playButton.setIcon(self.pause_icon)
         else:
             self.playButton.setIcon(self.play_icon)
@@ -182,7 +181,7 @@ class VideoWindow(QWidget):
         # print('VideoPlayer setGlobalPosition')
         if self.current_time_range[0] <= pos <= self.current_time_range[1]: # correct file opened
             position = int((pos-self.current_time_range[0])*1000)
-            if self.mediaPlayer.state() == QMediaPlayer.PlayingState and abs(position-self.last_position)<200:
+            if self.mediaPlayer.playbackState == QMediaPlayer.PlayingState and abs(position-self.last_position)<200:
                 # skip position setting by signal of main model to ensure smooth video plaback
                 return
             # go to correct relative position
@@ -201,19 +200,19 @@ class VideoWindow(QWidget):
                     self.errorLabel.setText("File: " + self.current_file)
                     self.current_time_range = arange
                     self.waiting_for_file = True
-                    self.media_state_before_file_transition = self.mediaPlayer.state()
+                    self.media_state_before_file_transition = self.mediaPlayer.playbackState
                     self.mediaPlayer.stop()
                     position = (pos-self.current_time_range[0])*1000
                     self.position_on_new_file = int(position)
                     # print('Changing position_on_new_file: ', self.position_on_new_file,pos)
-                    self.mediaPlayer.setMedia(QMediaContent(QUrl.fromLocalFile(file)))
+                    self.mediaPlayer.setMedia(QUrl.fromLocalFile(file))
                     self.playButton.setEnabled(True)
                     # self.duration = (arange[1]-arange[0])*1000
                     return
         logger.info(f'No video file found for current position {pos}')
         self.errorLabel.setText('No video file found for current position ' + str(pos))
         self.mediaPlayer.stop()
-        self.mediaPlayer.setMedia(QMediaContent())
+        self.mediaPlayer.setMedia()
         self.current_file = ''
         self.current_time_range = [0, 0]
         # self.duration = 0
